:truck: Move splash & loader to components folder
#92 splash, loader를 컴포넌트 폴더로 변경 및 그에 따른 각 페이지별 업데이트
