+++ src/components/Splash/Splash.jsx
