+++ src/components/Splash/Splash.styled.jsx
